requirements.txt
ai-tools==0.3.9
aiohttp
clip==1.0
cognitive-face==1.5.0
discord.py[voice]>=1.7.3
dlib==19.24.0
emoji==2.0.0
ffmpeg-python==0.2.0
jaconv==0.3
kornia==0.6.8
librosa==0.9.2
numpy==1.21.6
oauth2client
Pillow==9.2.0
pytorch-lightning==1.8.3.post1
pysndfile==1.4.4
rembg==2.0.25
requests
taming-transformers==0.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchmetrics==0.11.0
torchvision==0.14.0+cu117
zprint==0.0.11